@@ -1,28 +1,4 @@
-
-buildscript {
-    repositories {
-        jcenter()
-        google()
-    }
-
-    dependencies {
-        classpath 'com.android.tools.build:gradle:3.0.1'
-    }
-}
-
 apply plugin: 'com.android.application'
-
-repositories {
-    jcenter()
-    google()
-}
-
-dependencies {
-    compile "com.android.support:support-v4:27.0.2"
-    compile "com.android.support:support-v13:27.0.2"
-    compile "com.android.support:cardview-v7:27.0.2"
-    compile "com.android.support:appcompat-v7:27.0.2"
-}
 
 // The sample build uses multiple directories to
 // keep boilerplate and common code separate from
@@ -33,13 +9,13 @@ List<String> dirs = [
     'template'] // boilerplate code that is generated by the sample template process
 
 android {
-    compileSdkVersion 27
+    compileSdkVersion 29
 
-    buildToolsVersion "27.0.2"
+    buildToolsVersion "29.0.0"
 
     defaultConfig {
         minSdkVersion 18
-        targetSdkVersion 27
+        targetSdkVersion 29
     }
 
     compileOptions {
@@ -59,4 +35,9 @@ android {
 
     }
 
+}
+
+dependencies {
+    implementation 'androidx.appcompat:appcompat:1.0.2'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
 }
