@@ -33,7 +33,6 @@ package com.example.android.common.logger;
 
 import android.graphics.Typeface;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
 import android.text.Editable;
 import android.text.TextWatcher;
 import android.view.Gravity;
@@ -41,6 +40,8 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ScrollView;
+
+import androidx.fragment.app.Fragment;
 
 /**
  * Simple fraggment which contains a LogView and uses is to output log data it receives
